+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './common.js'
+export * from './image.js'
 export * from './item.js'
 export * from './order.js'
 export * from './question.js'